feat: modify services for user creation.
Modify the `adminService.js` and `userService.js` services for the
correct interfacing with the `admimModel.js` and `userModel.js` model
files.

With the added change that this removes a `readCSV` function that was
lingering there. But since it was empty, I guess it's fine to keep this
change in this rebasing.